pyproject.toml
[tool.poetry]
name = "whiskey"
version = "0.1.0"
description = ""
authors = ["Maximilian Linhoff <maximilian.linhoff@tu-dortmund.de>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
whitenoise = "^6.2.0"
djangorestframework = "^3.14.0"
rpi-gpio = "^0.7.1"
adafruit-circuitpython-neopixel = "^6.3.7"
rpi-ws281x = "^4.3.4"
redis = "^4.4.1"
daphne = "^4.0.0"
channels = "^4.0.0"
channels-redis = "^4.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
